Cambio de vistas para usuarios

+++ resources/views/usuarios/index.blade.php
@@ -5,12 +5,11 @@
 @stop
 @section ("contenido")
     <div class="spinner"></div>
-    <div class="hidden table-size-optima altoMaximo">
+    <div class="hidden table-size-optima">
         @if (Auth::check())
             @if (Auth::user()->roles->peso>=config('opciones.roles.administrador'))
-                <div class="row">
-                    @include('usuarios.Parciales.buscar')
-                </div>
+                @include('comun.plantillaBuscarIndex',['htmlTemplate'=>'usuarios.parciales.buscar'])
+                @include('comun.plantillaOperacionesIndex',['tabla'=>'usuarios','accion'=>'Nuevo'])
             @endif
             @if(!$users->isEmpty())
                 @foreach ($users as $usuario)
@@ -32,24 +31,35 @@
                                     </i>
                                 </a>
                                 @if (Auth::user()->roles->peso>=config('opciones.roles.administrador'))
-                                    {!! FORM::open(array('route' => array('usuarios.destroy',
-                                    $usuario->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
-                                    <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
-                                            @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                            class="pull-right lanzarModal"
-                                            data-title="BORRADO"
-                                            data-descripcion="¿Seguro que deseas eliminar este usuario?
+                                    @if($usuario->activo)
+                                        {!! FORM::open(array('route' => array('usuarios.destroy',
+                                        $usuario->id),'method' => 'DELETE','title'=>(config('opciones.accion.mostrarModalDeBorrado')?'':'Borrar'))) !!}
+                                        <button type="@if(config('opciones.accion.mostrarModalDeBorrado'))button @else submit @endif"
+                                                @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                                class="pull-right lanzarModal simpleModal"
+                                                data-modal_sin_etiqueta="true"
+                                                data-modal_ancho="330"
+                                                data-modal_cabecera_color_fondo='rgba(255,0,0,.9)'
+                                                data-modal_cabecera_color_texto='#ffffff'
+                                                data-modal_cuerpo_color_fondo='rgba(255,255,255,.9)'
+                                                data-modal_cuerpo_color_texto='"#ffffff'
+                                                data-modal_pie_color_fondo='#400090'
+                                                data-modal_pie_color_texto='"#ffffff'
+                                                data-modal_posicion_vertical="220"
+                                                data-titulo="BORRAR"
+                                                data-pie="true"
+                                                data-descripcion="¿Seguro que deseas eliminar este usuario?
                                                     <h3><strong class='green'>{{ $usuario->fullname}}</strong></h3>"
-                                            data-footer="true"
-                                            @endif >
-                                        <i class='glyphicon glyphicon-trash full-Width'>
-                                            <div>Borrar</div>
-                                        </i>
-                                    </button>
-                                    @if(config('opciones.accion.mostrarModalDeBorrado'))
-                                        @include ("comun.plantillaBorrado")
+                                                @endif >
+                                            <i class='glyphicon glyphicon-trash full-Width'>
+                                                <div>Borrar</div>
+                                            </i>
+                                        </button>
+                                        @if(config('opciones.accion.mostrarModalDeBorrado'))
+                                            @include ("comun.plantillaBorrado")
+                                        @endif
+                                        {!! FORM::close() !!}
                                     @endif
-                                    {!! FORM::close() !!}
                                 @endif
                             </th>
                         </tr>
@@ -96,19 +106,15 @@
                         </a>
                     </div>
                 @endif
+                {!! $users->appends(Request::only(['campo','value','rol']))->render()!!}
             @else
                 <div class="clearfix">
                     <div class="alert alert-info" role="alert">
-                        <p><strong>¡Aviso!</strong> No se ha encontrado ningún usuario que listar.</p>
+                        <p><strong>¡Aviso!</strong> No se ha encontrado ning&uacute;n usuario que listar.</p>
                     </div>
                 </div>
             @endif
-            @if (Auth::user()->roles->peso>=config('opciones.roles.administrador'))
-                    <div class="row paginationBlock">
-                    {!! $users->appends(Request::only(['campo','value','rol']))->render()
-                    !!}{{-- Poner el paginador --}}
-                </div>
-            @endif
+
         @else
             @include('comun.guestGoHome')
         @endif
